# openapi.yaml
openapi: 3.1.0
info:
  title: Afere API
  version: 3.1.0
  description: >
    Composition-builder API for the Afere medical procedure valuation platform.
    Implements CBHPM 2022 items 4.1, 4.2, and 5.1 for correct multi-procedure
    discounting and auxiliary surgeon remuneration.
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /api/calculations:
    post:
      summary: Persist a completed calculation
      description: >
        Saves a completed valuation to the database. The request must include a
        calculation_result with lead_surgeon_fee > 0 to ensure only real valuations
        are persisted. Returns the public_id for the shareable URL.
      operationId: saveCalculation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveCalculationRequest'
      responses:
        '201':
          description: Calculation saved. Returns the public_id for URL construction.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaveCalculationResponse'
        '400':
          description: Validation error (missing fields or incomplete valuation).
...
components:
  schemas:
...
    AccessRouteType:
      type: string
      enum: [same, different]
      description: >
        Whether all selected procedures were performed via the same access route
        (CBHPM 4.1 – 50% discount on secondary procedures) or different access
        routes (CBHPM 4.2 – 70% discount on secondary procedures).
    SelectedCode:
      type: object
      required: [cbhpm_code, description, porte]
      properties:
        cbhpm_code:
          type: string
          description: CBHPM code selected by the physician.
        description:
          type: string
          description: CBHPM description (passed through from the procedure detail).
        porte:
          type: string
          description: Porte for this code (from the SBN catalog, not editable by the physician).
...
    CodeBreakdown:
      type: object
      required: [cbhpm_code, description, porte, base_value, is_principal]
      properties:
        cbhpm_code:
          type: string
        description:
          type: string
        porte:
          type: string
        base_value:
          type: number
          format: float
          description: Full monetary value of this code's porte (before any discount).
        is_principal:
          type: boolean
          description: True for the highest-value code selected as the principal procedure.
    AuxiliaryFee:
      type: object
      required: [position, percentage, fee]
      properties:
        position:
          type: integer
          description: Auxiliary position (1 = first, 2 = second, etc.).
        percentage:
          type: number
          format: float
          description: Percentage applied (60, 40, 30, or 30).
        fee:
          type: number
          format: float
          description: Calculated fee for this auxiliary.
    SurgeonBreakdown:
      type: object
      required: [principal_value, additional_gross, discount_rate, additional_discounted, surgeon_total]
      properties:
        principal_value:
          type: number
          format: float
          description: Monetary value of the principal (highest-value) procedure.
        additional_gross:
          type: number
          format: float
          description: Sum of all non-principal procedure values before applying the CBHPM discount.
        discount_rate:
          type: number
          format: float
          description: Discount rate applied to additional procedures (0.50 for same route, 0.70 for different routes).
        additional_discounted:
          type: number
          format: float
          description: Sum of additional procedure values after applying the discount.
        surgeon_total:
          type: number
          format: float
          description: Final surgeon fee (principal_value + additional_discounted).
    CalculateResponse:
      type: object
      required:
        - code_breakdown
        - access_route_type
        - surgeon_breakdown
        - lead_surgeon_fee
        - individual_auxiliary_fees
        - auxiliaries_fee
        - anesthesiologist_fee
        - final_total
        - total_base
      properties:
        code_breakdown:
          type: array
          items:
            $ref: '#/components/schemas/CodeBreakdown'
          description: Per-code contribution with principal flag.
        access_route_type:
          $ref: '#/components/schemas/AccessRouteType'
        surgeon_breakdown:
          $ref: '#/components/schemas/SurgeonBreakdown'
        lead_surgeon_fee:
          type: number
          format: float
          description: Final surgeon fee (equals surgeon_breakdown.surgeon_total).
        individual_auxiliary_fees:
          type: array
          items:
            $ref: '#/components/schemas/AuxiliaryFee'
          description: Per-auxiliary fee details (CBHPM 5.1 – 60/40/30/30%).
        auxiliaries_fee:
          type: number
          format: float
          description: Sum of all individual auxiliary fees.
        anesthesiologist_fee:
          type: number
          format: float
          description: Fixed anesthesiologist fee (if applicable).
        final_total:
          type: number
          format: float
          description: Grand total (lead_surgeon_fee + auxiliaries_fee + anesthesiologist_fee).
        total_base:
          type: number
          format: float
          description: Simple sum of all selected code porte values (before discounting). Retained for compatibility.
    SaveCalculationRequest:
      type: object
      required: [procedure_name, selected_codes, auxiliaries_count, requires_anesthesia, access_route_type, calculation_result]
      properties:
        procedure_name:
          type: string
          minLength: 1
          description: SBN procedure name as displayed to the physician.
        procedure_sbn_code:
          type: string
          description: SBN catalog ID (optional; used for future cross-referencing).
        selected_codes:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/SelectedCode'
        auxiliaries_count:
          type: integer
          minimum: 0
          maximum: 4
        requires_anesthesia:
          type: boolean
        access_route_type:
          $ref: '#/components/schemas/AccessRouteType'
        calculation_result:
          $ref: '#/components/schemas/CalculateResponse'
          description: The completed valuation result. lead_surgeon_fee must be > 0.
    SaveCalculationResponse:
      type: object
      required: [public_id, created_at]
      description: >
        Only the public_id is returned. Use it to construct the shareable URL:
        https://afere.app/calc/{public_id}
      properties:
        public_id:
          type: string
          format: uuid
          description: UUID v4 public identifier for this calculation. Never the internal DB primary key.
        created_at:
          type: string
          format: date-time
...